+++ Sources/SwiftDoc/SourceFile.swift
@@ -56,14 +56,14 @@ public struct SourceFile: Hashable, Codable {
         func symbol<Node, Declaration>(_ type: Declaration.Type, _ node: Node) -> Symbol? where Declaration: API & ExpressibleBySyntax, Node == Declaration.Syntax, Node: SymbolDeclProtocol {
             guard let api = Declaration(node) else { return nil }
             guard let documentation = try? Documentation.parse(node.documentation) else { return nil }
-            let sourceLocation = sourceLocationConverter.location(for: node.position)
-            return Symbol(api: api, context: context, declaration: declaration(for: node), documentation: documentation, sourceLocation: sourceLocation)
+            let sourceRange = node.sourceRange(using: sourceLocationConverter)
+            return Symbol(api: api, context: context, declaration: declaration(for: node), documentation: documentation, sourceRange: sourceRange)
         }
 
         func symbol<Node: SymbolDeclProtocol>(_ node: Node, api: API) -> Symbol? {
             guard let documentation = try? Documentation.parse(node.documentation) else { return nil }
-            let sourceLocation = sourceLocationConverter.location(for: node.position)
-            return Symbol(api: api, context: context, declaration: declaration(for: node), documentation: documentation, sourceLocation: sourceLocation)
+            let sourceRange = node.sourceRange(using: sourceLocationConverter)
+            return Symbol(api: api, context: context, declaration: declaration(for: node), documentation: documentation, sourceRange: sourceRange)
         }
 
         func declaration<Node: SymbolDeclProtocol>(for node: Node) -> [Token] {